Create controller hanlers and custo error, catch exception on serive. Methods to cover test deleteShouldReturnNotFoundWhenNonExistingId

+++ src/main/java/com/devsuperior/bds02/services/CityService.java
@@ -12,6 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.devsuperior.bds02.dto.CityDTO;
 import com.devsuperior.bds02.entities.City;
 import com.devsuperior.bds02.repository.CityRepository;
+import com.devsuperior.bds02.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class CityService {
@@ -34,6 +35,9 @@ public class CityService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!cityRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource Not Found");
+        }
         cityRepository.deleteById(id);
     }
